test(types-1): triage facade suite to Spark answers
76 failures classified flip-or-revert: lawful flips pin the Spark
answer with citation; stale Spark-2.x-era no-engine tiers removed
from window parity; null-top last() corrected on both engines.
check_lib_py baselines amended (core.py +2, window parity ratchet).
pins: types-1/C-008
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

+++ python/repark/tests/test_iceberg_hygiene.py
@@ -117,7 +117,7 @@ def test_ref_ddl_create_branch_tag_time_travel(spark: ReparkSession) -> None:
 
     tag_arrow = spark.sql(f"SELECT id FROM {TABLE} VERSION AS OF 'tag_s1' ORDER BY id").to_arrow()
     assert _arrow_ids(tag_arrow) == [1, 2, 3]
-    assert _schema_names_types(tag_arrow) == [("id", "int64")]
+    assert _schema_names_types(tag_arrow) == [("id", "int32")]
 
     branch_arrow = spark.sql(
         f"SELECT id FROM {TABLE} VERSION AS OF 'branch_s2' ORDER BY id"
